feat: migrate engine to 64-bit architecture and optimize VM performance
- Expanded entity IDs and coordinates (X, Y, Z) to 64-bit `long` across all core models and interfaces.
- Refactored `DreamValue` with an explicit memory union for `double` and `long` to ensure absolute precision for 64-bit IDs.
- Overhauled `BytecodeInterpreter` using `ref struct` and `fixed` pointers for zero-overhead bytecode access.
- Inlined high-frequency math, trigonometric, string formatting, and control flow opcodes directly into the VM loop.
- Optimized `SpatialGrid` using `HashSet<IGameObject>` and `ReaderWriterLockSlim` to support high-density entity management and concurrent queries.
- Streamlined `GameState` for high-throughput entity addition/removal and state change tracking.
- Verified stability with 180 unit tests and a new 100,000-object stress test suite.

+++ Engine/Core/LuaGameApi.cs
@@ -14,11 +14,11 @@ namespace Core
         }
 
         public IMap? GetMap() => _gameApi.Map.GetMap();
-        public ITurf? GetTurf(int x, int y, int z) => _gameApi.Map.GetTurf(x, y, z);
-        public void SetTurf(int x, int y, int z, int turfId) => _gameApi.Map.SetTurf(x, y, z, turfId);
-        public GameObject? CreateObject(int typeId, int x, int y, int z) => _gameApi.Objects.CreateObject(typeId, x, y, z);
-        public GameObject? GetObject(int id) => _gameApi.Objects.GetObject(id);
-        public void DestroyObject(int id) => _gameApi.Objects.DestroyObject(id);
+        public ITurf? GetTurf(long x, long y, long z) => _gameApi.Map.GetTurf(x, y, z);
+        public void SetTurf(long x, long y, long z, int turfId) => _gameApi.Map.SetTurf(x, y, z, turfId);
+        public GameObject? CreateObject(int typeId, long x, long y, long z) => _gameApi.Objects.CreateObject(typeId, x, y, z);
+        public GameObject? GetObject(long id) => _gameApi.Objects.GetObject(id);
+        public void DestroyObject(long id) => _gameApi.Objects.DestroyObject(id);
 
         public void LoadMap(string filePath)
         {
